# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NYX API
  version: "1.0.0"
servers:
  - url: /api
paths:
...
  /v1/approvals:
    get:
      summary: List approvals
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - in: query
          name: after
          schema:
            type: string
      responses:
        "200":
          description: Paginated approval list
  /v1/approvals/{approvalId}:
    get:
      summary: Get approval detail
      parameters:
        - in: path
          name: approvalId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approval payload
...